pyproject.toml
[tool.poetry]
name = "rpa-challenge"
version = "0.1.0"
description = ""
authors = ["felipejazz <aluvinof@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
rpaframework = "^28.6.1"
openpyxl = "^3.1.5"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
